+++ canaille/core/forms.py
@@ -285,10 +285,22 @@ def profile_form(write_field_names, readonly_field_names, user=None):
     if "groups" in fields and not models.Group.query():
         del fields["groups"]
 
+    fields["lock_date"] = wtforms.DateTimeLocalField(
+        _("Account expiration"),
+        validators=[wtforms.validators.Optional()],
+        format=[
+            "%Y-%m-%d %H:%M",
+            "%Y-%m-%dT%H:%M",
+            "%Y-%m-%d %H:%M:%S",
+            "%Y-%m-%dT%H:%M:%S",
+        ],
+    )
+
     form = HTMXBaseForm(fields)
     form.user = user
     for field in form:
         if field.name in readonly_field_names - write_field_names:
+            field.render_kw = field.render_kw or {}
             field.render_kw["readonly"] = "true"
 
     return form