+++ tests/core/test_account.py
@@ -1,3 +1,4 @@
+import datetime
 from unittest import mock
 
 from canaille.app import models
@@ -85,6 +86,21 @@ def test_signin_wrong_password(testclient, user):
     res = res.form.submit(status=302)
     res = res.follow(status=200)
     res.form["password"] = "incorrect horse"
+    res = res.form.submit(status=200)
+    assert ("error", "Login failed, please check your information") in res.flashes
+
+
+def test_signin_bad_csrf(testclient, user):
+    with testclient.session_transaction() as session:
+        assert not session.get("user_id")
+
+    res = testclient.get("/login", status=200)
+
+    res.form["login"] = "John (johnny) Doe"
+    res = res.form.submit(status=302)
+    res = res.follow(status=200)
+
+    res.form["password"] = ""
     res = res.form.submit(status=200)
     assert ("error", "Login failed, please check your information") in res.flashes
 
@@ -336,3 +352,94 @@ def test_user_self_deletion(testclient, backend):
         assert not sess.get("user_id")
 
     testclient.app.config["ACL"]["DEFAULT"]["PERMISSIONS"] = []
+
+
+def test_account_locking(user, backend):
+    assert not user.locked
+    assert not user.lock_date
+    assert user.check_password("correct horse battery staple") == (True, None)
+
+    user.lock()
+    user.save()
+    assert user.locked
+    assert user.lock_date <= datetime.datetime.now(datetime.timezone.utc)
+    assert models.User.get(id=user.id).locked
+    assert models.User.get(id=user.id).lock_date <= datetime.datetime.now(
+        datetime.timezone.utc
+    )
+    assert user.check_password("correct horse battery staple") == (
+        False,
+        "Your account has been locked.",
+    )
+
+    user.unlock()
+    user.save()
+    assert not user.locked
+    assert not user.lock_date
+    assert not models.User.get(id=user.id).locked
+    assert not models.User.get(id=user.id).lock_date
+    assert user.check_password("correct horse battery staple") == (True, None)
+
+
+def test_account_locking_past_date(user, backend):
+    assert not user.locked
+    assert not user.lock_date
+    assert user.check_password("correct horse battery staple") == (True, None)
+
+    lock_datetime = datetime.datetime.now(datetime.timezone.utc).replace(
+        microsecond=0
+    ) - datetime.timedelta(days=30)
+    user.lock(lock_datetime)
+    user.save()
+    assert user.locked
+    assert user.lock_date == lock_datetime
+    assert models.User.get(id=user.id).locked
+    assert models.User.get(id=user.id).lock_date == lock_datetime
+    assert user.check_password("correct horse battery staple") == (
+        False,
+        "Your account has been locked.",
+    )
+
+
+def test_account_locking_future_date(user, backend):
+    assert not user.locked
+    assert not user.lock_date
+    assert user.check_password("correct horse battery staple") == (True, None)
+
+    lock_datetime = datetime.datetime.now(datetime.timezone.utc).replace(
+        microsecond=0
+    ) + datetime.timedelta(days=365 * 4)
+    user.lock(lock_datetime)
+    user.save()
+    assert not user.locked
+    assert user.lock_date == lock_datetime
+    assert not models.User.get(id=user.id).locked
+    assert models.User.get(id=user.id).lock_date == lock_datetime
+    assert user.check_password("correct horse battery staple") == (True, None)
+
+
+def test_signin_locked_account(testclient, user):
+    with testclient.session_transaction() as session:
+        assert not session.get("user_id")
+
+    user.lock()
+    user.save()
+
+    res = testclient.get("/login", status=200)
+    res.form["login"] = "user"
+
+    res = res.form.submit(status=302).follow(status=200)
+    res.form["password"] = "correct horse battery staple"
+
+    res = res.form.submit()
+    res.mustcontain("Your account has been locked.")
+
+    user.unlock()
+    user.save()
+
+
+def test_account_locked_during_session(testclient, logged_user):
+    testclient.get("/profile/user/settings", status=200)
+    logged_user.lock()
+    logged_user.save()
+    testclient.get("/profile/user/settings", status=403)