+++ tests/oidc/test_password_flow.py
@@ -63,12 +63,31 @@ def test_password_flow_post(testclient, user, client):
     assert res.json["name"] == "John (johnny) Doe"
 
 
-def test_invalid_credentials(testclient, user, client):
+def test_invalid_user(testclient, user, client):
     res = testclient.post(
         "/oauth/token",
         params=dict(
             grant_type="password",
             username="invalid",
+            password="invalid",
+            scope="openid profile groups",
+        ),
+        headers={"Authorization": f"Basic {client_credentials(client)}"},
+        status=400,
+    )
+
+    assert res.json == {
+        "error": "invalid_request",
+        "error_description": 'Invalid "username" or "password" in request.',
+    }
+
+
+def test_invalid_credentials(testclient, user, client):
+    res = testclient.post(
+        "/oauth/token",
+        params=dict(
+            grant_type="password",
+            username="user",
             password="invalid",
             scope="openid profile groups",
         ),