+++ CHANGES.rst
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`_,
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
+Added
+*****
+
+- Implemented account expiration based on OpenLDAP ppolicy overlay. Needs OpenLDAP 2.5+
+  :issue:`13` :pr:`118`
+
 Fixed
 *****